график по длительности
