+++ package.json
@@ -25,7 +25,7 @@
   "dependencies": {
     "chart.js": "2.9.4",
     "chartjs-plugin-datalabels": "0.7.0",
-    "dayjs": "1.9.5",
+    "dayjs": "1.9.7",
     "flatpickr": "4.6.6"
   }
 }
