+++ src/utils/common.js
@@ -1,4 +1,7 @@
 import dayjs from 'dayjs';
+import {TYPES} from '../const.js';
+import duration from 'dayjs/plugin/duration';
+dayjs.extend(duration);
 import {RANDOM_MIN_DAY, RANDOM_MAX_DAY, RANDOM_MIN_TIME, RANDOM_MAX_TIME} from '../const.js';
 /**
  * Функция из интернета по генерации случайного числа из диапазона
@@ -57,4 +60,17 @@ export const arrayEqual = (arrayA, arrayB) => {
       return false;
 
   return true;
+};
+
+export const durationFormat = (val) => {
+  const durationDaysTemplate = (dayjs.duration(val, 'minutes').days()==0) ? '' : dayjs.duration(val,'minutes').days() + 'D ';
+  const durationHoursTemplate = dayjs.duration(val,'minutes').format('HH') + 'H ';
+  const durationMinutesTemplate = dayjs.duration(val,'minutes').format('MM') + 'M';
+  return durationDaysTemplate + durationHoursTemplate + durationMinutesTemplate;
+};
+
+export const getPointTypesInUpperCase = () => {
+  const pointTypes = [];
+  TYPES.forEach((element) => pointTypes.push(element.name.toUpperCase()));
+  return pointTypes;
 };
